+++ src/skin.js
@@ -17,31 +17,17 @@ function Skin(FS) {
 
         dom.nodeAddClass(FS.root, "fsjs-skin")
 
+
         const rangeInputs = FS.root.querySelectorAll("input[type=range][data-fsjs-ui='slider']")
         if (rangeInputs.length) {
-            // rangeSlider.create(rangeInputs, {
-            //     polyfill: true,
-            //     // utilise the more granular events offered by the skin
-            //     // default html range inputs only trigger on change
-            //     onSlide: updateSlider,
-            //     onSlideEnd: updateSlider
-            // })
             rangeInputs.forEach(slider => {
                 let mouseDown = false;
 
-                slider.addEventListener("mousedown", function () {
-                    mouseDown = true
-                })
-
-                slider.addEventListener("mouseup", function () {
-                    mouseDown = false
-                })
-
-                slider.addEventListener("mousemove", function (e) {
+                function updateSlider (e) {
                     if (!mouseDown) {
                         return
                     }
-
+        
                     const w = slider.getBoundingClientRect().width,
                         x = Math.max(0, Math.min(w, e.layerX)),
                         percent = x / w,
@@ -50,26 +36,37 @@ function Skin(FS) {
                         step = parseFloat(slider.step) || 1, // _some_ defaults
                         step_decimals = utils.countDecimals(slider.step),
                         range = min < max ? max - min : min - max;
-
+        
                     let key = slider.dataset.fsjs,
                         value;
-
+        
                     value = Math.round((min + (percent * range)) / step) * step;
                     
                     if (step_decimals > 0) {
                         value = value.toPrecision(step_decimals)
                     }
-
-                    // Catch special case for variable font axis sliders
+        
                     if (typeof (key) === "undefined") {
+                        // Set variable axis slider
                         let key = slider.dataset.axis;
                         opt[key] = value
                         FS.setValue("variation", opt)
                     } else {
+                        // Set regular slider (leading, letter spacing, ...)
                         FS.setValue(key, value)
                     }
+                }
 
+                slider.addEventListener("mousedown", function (e) {
+                    mouseDown = true
+                    updateSlider(e)
                 })
+
+                slider.addEventListener("mouseup", function () {
+                    mouseDown = false
+                })
+
+                slider.addEventListener("mousemove", updateSlider)
             })
         }